+++ src/components/AppFooter.vue
@@ -61,7 +61,7 @@
         <div style="font-size: 12px">
           {{ new Date().getFullYear() }}
           <strong>
-            <a href="http://www.icjia.state.il.us" class="footer-link"
+            <a href="https://www.icjia.state.il.us" class="footer-link"
               >Illinois Criminal Justice Information Authority</a
             > </strong
           >&nbsp;|&nbsp;