+++ src/main.js
@@ -1,4 +1,3 @@
-import "babel-polyfill";
 import Vue from "vue";
 import App from "./App.vue";
 import router from "./router";
@@ -6,20 +5,6 @@ import store from "./store";
 import vuetify from "@/plugins/vuetify";
 import "@/filters";
 import "@/css/app.css";
-
-// import AOS from "aos";
-// import "aos/dist/aos.css";
-
-(function () {
-  if (typeof NodeList.prototype.forEach === "function") return false;
-  NodeList.prototype.forEach = Array.prototype.forEach;
-})();
-
-(function () {
-  if (!Element.prototype.matches) {
-    Element.prototype.matches = Element.prototype.msMatchesSelector;
-  }
-})();
 
 import browserDetect from "vue-browser-detect-plugin";
 Vue.use(browserDetect);